requirements/local.txt
-r base.txt

gunicorn==20.1.0  # https://github.com/benoitc/gunicorn

# Django
# ------------------------------------------------------------------------------
django-storages[boto3]==1.13.1  # https://github.com/jschneier/django-storages
django-anymail[amazon_ses]==9.0  # https://github.com/anymail/django-anymail

# Code quality
# ------------------------------------------------------------------------------
black==22.12.0  # https://github.com/psf/black
factory-boy==3.3.1
django-extensions==3.2.3